+++ app/javascript/wilday_ui/controllers/index.js
@@ -1,7 +1,7 @@
 import { Application } from "@hotwired/stimulus";
 import ButtonController from "./button_controller";
 import DropdownController from "./dropdown_controller";
-
+import ClipboardController from "./clipboard_controller";
 // Initialize Stimulus
 const application = Application.start();
 window.Stimulus = application;
@@ -9,6 +9,7 @@ window.Stimulus = application;
 // Register the button controller
 application.register("button", ButtonController);
 application.register("dropdown", DropdownController);
+application.register("clipboard", ClipboardController);
 // Debug check to ensure Stimulus is loaded
 // if (window.Stimulus) {
 //   console.log("✅ Stimulus is loaded and initialized.");