+++ test/dummy/app/controllers/components/button/buttons_controller.rb
@@ -36,6 +36,10 @@ module Components
       def gradient
         render "components/button/gradient"
       end
+
+      def clipboard
+        render "components/button/clipboard"
+      end
     end
   end
 end