+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [TO BE RELEASED]
+- Added `copy_to_clipboard` feature for `w_button` component to copy text to clipboard with a feedback message.
+
 ## [0.6.0] - 2024-12-18
 
 - Added `:gradient` parameter for `w_button` component to use gradient color.